Breakup tests and add delete import call
