+++ tests/api_url.py
@@ -1,0 +1,2 @@
+def api_url(path: str):
+    return "https://api.foxglove.dev" + path
